requirements.txt
flask==3.0.3
flask-cors==4.0.1
gspread==6.1.2
google-auth==2.29.0
apscheduler==3.10.4
gunicorn==22.0.0
python-dateutil==2.9.0
